[PlaybackSerialiser] Rewrite DesyncMonitor as file command extension
- Move DesyncMonitoring to DesyncMonitorFileCommandExtension
- Move Monitor to Utils
- Remove seed value from DesyncMonitor
- Registering new FileCommand in TASmodClient
