+++ src/main/java/com/minecrafttas/tasmod/util/Monitor.java
@@ -1,4 +1,4 @@
-package com.minecrafttas.tasmod.monitoring;
+package com.minecrafttas.tasmod.util;
 
 import java.lang.reflect.Field;
